Delete /set messages once the device has processed it
Note: the logic is very simple, after the merged message is sent a flag is set. As soon as a device sends status, it is cleared (and the topic deleted). In reality we should actually check the merged message was acknowledged by esp_now, and then set the flag.

Also, separate pairing time-out and data-timeout,